+++ hawc/apps/epiv2/exports.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from django.db.models import CharField, F, Func, QuerySet, Value
 
-from ...services.excel import get_writer, write_worksheet
+from ...tools.excel import get_writer, write_worksheet
 from ..common.exports import Exporter, ModelExport
 from ..common.helper import FlatFileExporter
 from ..common.models import sql_display, sql_format, str_m2m, to_display_array
